Embedding more data into events

+++ src/core/GameScreen.h
@@ -59,7 +59,7 @@ public:
     
     int getPlayerDirection();
     
-    short popOldestEventId();
+    int popOldestEventId();
     
     bool isTimeToSendKeepAlive();
